+++ src/commands/showCommitSearch.ts
@@ -12,8 +12,8 @@ import { ShowQuickCommitDetailsCommandArgs } from './showQuickCommitDetails';
 const searchByRegex = /^([@~=:#])/;
 const searchByMap = new Map<string, GitRepoSearchBy>([
     ['@', GitRepoSearchBy.Author],
-    ['~', GitRepoSearchBy.Changes],
-    ['=', GitRepoSearchBy.ChangedOccurrences],
+    ['~', GitRepoSearchBy.ChangedLines],
+    ['=', GitRepoSearchBy.Changes],
     [':', GitRepoSearchBy.Files],
     ['#', GitRepoSearchBy.Sha]
 ]);
@@ -65,7 +65,7 @@ export class ShowCommitSearchCommand extends ActiveEditorCachedCommand {
             args.search = await window.showInputBox({
                 value: args.search,
                 prompt: `Please enter a search string`,
-                placeHolder: `search by message, author (@<pattern>), files (:<pattern>), commit id (#<sha>), changes (~<pattern>), or changed occurrences (=<string>)`
+                placeHolder: `search by message, author (@<pattern>), files (:<pattern>), commit id (#<sha>), changes (=<pattern>), changed lines (~<pattern>)`
             } as InputBoxOptions);
             if (args.search === undefined) {
                 return args.goBackCommand === undefined ? undefined : args.goBackCommand.execute();
@@ -96,12 +96,12 @@ export class ShowCommitSearchCommand extends ActiveEditorCachedCommand {
                 searchLabel = `commits with an author matching '${args.search}'`;
                 break;
 
-            case GitRepoSearchBy.Changes:
-                searchLabel = `commits with changes matching '${args.search}'`;
+            case GitRepoSearchBy.ChangedLines:
+                searchLabel = `commits with changed lines matching '${args.search}'`;
                 break;
 
-            case GitRepoSearchBy.ChangedOccurrences:
-                searchLabel = `commits with changed occurrences matching '${args.search}'`;
+            case GitRepoSearchBy.Changes:
+                searchLabel = `commits with changes matching '${args.search}'`;
                 break;
 
             case GitRepoSearchBy.Files: